% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dodata.R
\name{Dodata}
\alias{Dodata}
\title{To perform method}
\usage{
Dodata(dir, calculatedata, chr, color1, CLO)
}
\arguments{
\item{dir}{the path of the output}

\item{calculatedata}{the inputdata}

\item{chr}{chromosome}

\item{color1}{the color of the chromosome}

\item{CLO}{the numbers of CPU}
}
\value{
list
}
\description{
To perform method
}
\examples{
data(BSA)
dir<- tempdir()
data.calculatedata<-Readdata1(BSA)
Dodata(dir,calculatedata=data.calculatedata,chr="all",color1="blue",CLO=1)
}
