% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dQTGseq.R
\name{dQTGseq}
\alias{dQTGseq}
\title{Title The function of dQTG.seq}
\usage{
dQTGseq(dir, filegen, chr, color, CLO)
}
\arguments{
\item{dir}{the path of the output}

\item{filegen}{the input data}

\item{chr}{the chromosome}

\item{color}{the color}

\item{CLO}{the numbers of CPU}
}
\value{
list
}
\description{
Title The function of dQTG.seq
}
\examples{
data(BSA)
dQTGseq(dir=tempdir(),filegen=BSA,chr="all",color="blue",CLO=1)
}
