% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{Readdata1}
\alias{Readdata1}
\title{Title readdata function}
\usage{
Readdata1(File)
}
\arguments{
\item{File}{the input file}
}
\value{
list
}
\description{
Title readdata function
}
\examples{
data(BSA)
Readdata1(BSA)
}
